Split StorageList file
